feat(repository): Update ActivityRepository to use NewVantage column names
- Replace all old column names with NewVantage names (ProjectID vs Tag_ProjectID)
- Update SQL queries to use new column names directly
- Fix PercentEntry calculation to use 0-100 format (divide by 100 for calculations)
- Update data reading to match new column names
- Remove dependency on ColumnMapper.GetDbColumnName() in repository layer
- Fix EarnedMHsRoc and ROCID data types (REAL instead of INTEGER)

BREAKING CHANGE: Repository now expects database with NewVantage column names.
Must delete old database before running.

Part of Phase 2 & 3: Activity Model & Repository Refactor
Ref: REFACTOR_PLAN.md

+++ DatabaseSetup.cs
@@ -97,10 +97,10 @@ namespace VANTAGE
       BudgetHoursGroup REAL DEFAULT 0,
        BudgetHoursROC REAL DEFAULT 0,
  BaseUnit REAL DEFAULT 0,
-         EarnedMHsRoc INTEGER DEFAULT 0,
+         EarnedMHsRoc REAL DEFAULT 0,
        
         -- ROC
-         ROCID INTEGER DEFAULT 0,
+         ROCID REAL DEFAULT 0,
        ROCPercent REAL DEFAULT 0,
            ROCBudgetQTY REAL DEFAULT 0,
